Add test-jar functionality, options for JAR tasks
Include the ability to add the tests as a JAR similar to the test-jar goal in maven,
as well as providing the ability to optionally specify which of sources, javadoc and
tests should be included in the list of artifacts to upload to a Nexus repository.

Defaults in the convention have changed to _not_ include any of the three by default;
while this changes the default previous behavior, it seems appropriate to opt-in, not
opt-out, on a per-project basis.

Minor: correct camel-casing of 'JavaDoc' to reflect common case rules, make consistent
the references internally.